pyproject.toml
[project]
name = "percival-discord-bot"
version = "0.1.0"
description = "An AI-powered Discord bot that uses OpenRouter to answer questions."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]
